feat: DBManager 클래스 추가
* DBManager.java: 데이터베이스 연결 및 해제를 담당하는 클래스
