+++ src/main/java/lotto/dao/LottoDAO.java
@@ -9,47 +9,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class LottoDAO {
-    public Connection getConnection() {
-        Connection con = null;
-        String server = "localhost:3307"; // MySQL 서버 주소
-        String database = "lotto_game"; // MySQL DATABASE 이름
-        String userName = "codemcd"; //  MySQL 서버 아이디
-        String password = "0803"; // MySQL 서버 비밀번호
+    private Connection con;
 
-        // 드라이버 로딩
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        } catch (ClassNotFoundException e) {
-            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
-            e.printStackTrace();
-        }
-
-        // 드라이버 연결
-        try {
-            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database +
-                    "?serverTimezone=UTC&allowPublicKeyRetrieval=true&useSSL=false", userName, password);
-            System.out.println("정상적으로 연결되었습니다.");
-        } catch (SQLException e) {
-            System.err.println("연결 오류:" + e.getMessage());
-            e.printStackTrace();
-        }
-
-        return con;
-    }
-
-    // 드라이버 연결해제
-    public void closeConnection(Connection con) {
-        try {
-            if (con != null)
-                con.close();
-        } catch (SQLException e) {
-            System.err.println("con 오류:" + e.getMessage());
-        }
+    public LottoDAO(Connection con) {
+        this.con = con;
     }
 
     public void addLottos(String lottoId, Lottos lottos) throws SQLException {
         String query = "INSERT INTO lotto VALUES (?, ?)";
-        PreparedStatement pstmt = getConnection().prepareStatement(query);
+        PreparedStatement pstmt = con.prepareStatement(query);
         pstmt.setString(1, lottoId);
         StringBuilder stringLottos = new StringBuilder();
         List<String> strings = new ArrayList<>();
@@ -63,7 +31,7 @@ public class LottoDAO {
 
     public Lottos findByLottoId(String lottoId) throws SQLException {
         String query = "SELECT * FROM lotto WHERE lotto_id = ?";
-        PreparedStatement pstmt = getConnection().prepareStatement(query);
+        PreparedStatement pstmt = con.prepareStatement(query);
         pstmt.setString(1, lottoId);
         ResultSet rs = pstmt.executeQuery();
